+++ Assets/Scripts/Turret.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 public class Turret : MonoBehaviour, IBuyable, IUpgradable
 {
-    public float costBuy;
-    public float costUpgrade;
+    public int costBuy;
+    public int costUpgrade;
 
     public bool Upgraded { get; set;}
